Move the rotateImage from the icon bar to text
